source/docker_app/requirements.txt
streamlit==1.29.0
boto3==1.34.125
streamlit-cognito-auth==1.3.1
pandas==2.2.2
matplotlib==3.9.0
aws-cdk-lib==2.146.0
langchain-anthropic==0.1.15
langchain-core== 0.2.7
langchain-aws==0.1.6
langchain-community==0.2.5
langsmith==0.1.77
langgraph==0.0.69
numpy==1.26.4
requests==2.32.3
pypdf==4.2.0
PyPDF2==3.0.1
#psycopg2==2.9.9
psycopg2-binary==2.9.9
psycopg==3.1.19
psycopg-binary==3.1.19
yfinance==0.2.40
pandas-market-calendars==4.4.1
langchain-postgres==0.0.9
python-docx==1.1.2
peewee==3.17.5
auth
